% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateCumulativeWater.R
\name{CalculateCumulativeWaterDepth}
\alias{CalculateCumulativeWaterDepth}
\title{Calculate Available Water from Individual Sensor Measurements}
\usage{
CalculateCumulativeWaterDepth(
  data = NA,
  design.col = NA,
  time.col = NA,
  depth.col = NA,
  response.col = NA,
  profile.min = NA,
  profile.max = NA,
  increment = 5,
  sensor.height = 10,
  divideby100 = T
)
}
\arguments{
\item{data}{Data frame. Must contain at least 3 columns with one timestamp column in POSIXct/POSIXlt format, one column to indicate the depth at which the sensors measuring volumetric water content were installed, and one column with the measured volumetric water content data in the numeric format. Additional columns can be added to characterize experimental design.}

\item{design.col}{List the induces of the optional columns characterizing the experimental design.}

\item{time.col}{Index of the timestamp column.}

\item{depth.col}{Index of the column containing the provided measured volumetric water content data.}

\item{response.col}{Index of the column indicating the depth at which the sensors measuring volumetric water content were installed.}

\item{profile.min}{Lower depth of the soil profile.}

\item{profile.max}{Upper depth of the soil profile.}

\item{increment}{Depth of the soil profile layers used to calculate cumulative depths of water in the profile.}

\item{sensor.height}{Thickness of soil affecting the volumetric soil water content measurement.}

\item{divideby100}{Equals T if the volumetric water content values are provided in percent rather than decimal values.}
}
\value{
A data frame with the coefficients used to calculate the weighted average of volumetric water content, a data frame with calculate cumulative depth of water.
}
\description{
This function calculate cumulative water depth from individual volumetric soil water content measurements
}
