% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DivideDataIntoChunks.R
\name{DivideIntoChunks}
\alias{DivideIntoChunks}
\title{Divide Cumulative Water Depth Data into Chunks of Continuous Data}
\usage{
DivideIntoChunks(
  data = NA,
  design.col = NA,
  time.col = NA,
  response.col = NA,
  min.successive.obs = 12,
  time.interval = 1,
  time.unit = "hour"
)
}
\arguments{
\item{data}{Data frame. Must contain at least 2 columns with one timestamp column in POSIXct/POSIXlt format and one column with the calculated depths of water. Additional columns can be added to characterize experimental design.}

\item{design.col}{List the induces of the optional columns characterizing the experimental design.}

\item{time.col}{Index of the timestamp column.}

\item{response.col}{Index of the column indicating the depth at which the sensors measuring volumetric water content were installed.}

\item{min.successive.obs}{Minimum number of observations in a chunk.}

\item{time.interval}{Temporal resolution.}

\item{time.unit}{Temporal unit.}
}
\value{
A data frame with the coefficients used to calculate the weighted average of volumetric water content, a data frame with calculate cumulative depth of water.
}
\description{
This function divides the dataframes of available water into chunks of continuous data
}
