% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemoveHighFrequencyNoise.R
\name{FilterWCData}
\alias{FilterWCData}
\title{Filter Data for High Frequency Noise}
\usage{
FilterWCData(data = NA, response.col = NA, nb.gs.filters = 2)
}
\arguments{
\item{data}{List of data divided into chunks.}

\item{response.col}{Index of the column indicating the depth at which the sensors measuring volumetric water content were installed.}

\item{nb.gs.filters}{Number of consecutive Savitsky-Golay filters applied to the data.}
}
\value{
A data frame with the coefficients used to calculate the weighted average of volumetric water content, a data frame with calculate cumulative depth of water.
}
\description{
This functions filters the calculated depth of water using one or multiple consecutive Savitsky-Golay filters.
}
