% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentifyInfiltrationEvents.R
\name{IdentifyInfiltrationEvents}
\alias{IdentifyInfiltrationEvents}
\title{Identify Infiltration Event within Available Water Data}
\usage{
IdentifyInfiltrationEvents(
  data = NA,
  time.col = NA,
  design.col = NA,
  response.col = NA,
  butter.n = 2,
  butter.w = 0.075,
  sensitivity = 0.1,
  approx.delay = 8,
  temp.resolution = 1,
  temp.unit = "hour"
)
}
\arguments{
\item{data}{List of data divided into chunks.}

\item{time.col}{Index of the timestamp column.}

\item{design.col}{List the induces of the optional columns characterizing the experimental design.}

\item{response.col}{Index of the column indicating the depth at which the sensors measuring volumetric water content were installed.}

\item{butter.n}{Butterworth filter order.}

\item{butter.w}{Critical frequency of the butterworth filter. Smaller w values increase the sensitivity of the created algorithm.}

\item{sensitivity}{Relative sensitivity of the algorithm. Lower values increase the sensitivity of the created algorithm.}

\item{approx.delay}{Estimated delay between the provided and filtered data. Default is 8 hours. Can be adjusted to improve algorithm performance.}

\item{temp.resolution}{Temporal resolution of the provided data, numeric value.}

\item{temp.unit}{Temporal resolution of the provided data, unit. Should be equal to "hour" or "min".}
}
\value{
A data frame with interpolated missing volumetric water content.
}
\description{
This function identifies infiltration events within calculated available water data.
}
