% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterpolateMissingVWCData.R
\name{InterpolateMissingVWCData}
\alias{InterpolateMissingVWCData}
\title{Interpolate Missing Volumetric Water Content Data}
\usage{
InterpolateMissingVWCData(
  data = NA,
  temp.resolution = 1,
  temp.unit = "hours",
  time.col = NA,
  response.col = NA,
  design.col = NA,
  depth.col = NA,
  depths.values = c(-15, -45, -80),
  missing.max.time.by.depth = c(12, NA, NA)
)
}
\arguments{
\item{data}{Data frame. Must contain at least 3 columns with one timestamp column in POSIXct/POSIXlt format, one column to indicate the depth at which the sensors measuring volumetric water content were installed, and one column with the measured volumetric water content data in the numeric format. Additional columns can be added to characterize experimental design.}

\item{temp.resolution}{Temporal resolution of the provided data, numeric value.}

\item{temp.unit}{Temporal resolution of the provided data, unit. Should be equal to "hour" or "min".}

\item{time.col}{Index of the timestamp column.}

\item{response.col}{Index of the column indicating the depth at which the sensors measuring volumetric water content were installed.}

\item{design.col}{List the induces of the optional columns characterizing the experimental design.}

\item{depth.col}{Index of the column containing the provided measured volumetric water content data.}

\item{depths.values}{List the depth values at which the sensors measuring volumetric water content were installed.}

\item{missing.max.time.by.depth}{List the maximum amount of missing time allowed for interpolation. Position of the element in the list corresponds to the corresponding depth in depths.values. Values are provided in number of temp.resolution intervals. Set value to NA if data should be interpolated no matter the number of missing observations.}
}
\value{
A data frame with interpolated missing volumetric water content.
}
\description{
This function interpolates missing volumetric water content data according to the user's specification. Missing volumetric water content values are estimated using a linear interpolation.
}
